+++ src-elevated-sidecar/src/main.rs
@@ -6,36 +6,74 @@
 extern crate lazy_static;
 use hyper::service::{make_service_fn, service_fn};
 use hyper::Server;
+use log::{error, info};
 use oyasumi_shared::models::ElevatedSidecarInitRequest;
 use std::convert::Infallible;
 use std::env;
+use std::fs::{
+    // OpenOptions, 
+    File
+};
 use std::net::SocketAddr;
+use std::path::Path;
 use std::time::Duration;
 use sysinfo::{Pid, PidExt, System, SystemExt};
 use windows::{is_elevated, relaunch_with_elevation};
+mod afterburner;
 mod http_handler;
 mod nvml;
 mod windows;
+use directories::BaseDirs;
+use simplelog::{
+    ColorChoice, CombinedLogger, Config, LevelFilter, TermLogger, TerminalMode, WriteLogger,
+};
 
 #[tokio::main]
 async fn main() {
+    // Initialize logging
+    let log_path = if let Some(base_dirs) = BaseDirs::new() {
+        base_dirs
+            .preference_dir()
+            .join("co.raphii.oyasumi/logs/OyasumiSidecar.log")
+    } else {
+        Path::new("co.raphii.oyasumi/logs/OyasumiSidecar.log").to_path_buf()
+    };
+    CombinedLogger::init(vec![
+        TermLogger::new(
+            LevelFilter::Info,
+            Config::default(),
+            TerminalMode::Mixed,
+            ColorChoice::Auto,
+        ),
+        WriteLogger::new(
+            LevelFilter::Info,
+            Config::default(),
+            File::create(log_path).unwrap(),
+            // OpenOptions::new()
+            //     .create(true) 
+            //     .append(true)
+            //     .open(log_path)
+            //     .unwrap(),
+        ),
+    ])
+    .unwrap();
     // Get port of host http server from 1st argument
     let args: Vec<String> = env::args().collect();
     if args.len() < 3 {
-        println!("Missing arguments. Expected format");
-        println!("oyasumi-admin.exe [main-http-port] [main-process-id]");
+        error!("Missing arguments. Expected format");
+        error!("oyasumi-admin.exe [main-http-port] [main-process-id]");
         std::process::exit(0);
     }
     let host_port = if let Ok(n) = args[1].parse::<u16>() {
         n
     } else {
-        println!("Invalid port number");
+        error!("Invalid port number");
         std::process::exit(0);
     };
     let main_pid = if let Ok(n) = args[2].parse::<u32>() {
         n
     } else {
-        println!("Invalid main process id");
+        error!("Invalid main process id");
         std::process::exit(0);
     };
     // Relaunch as admin if not elevated
@@ -65,7 +103,7 @@ async fn main() {
         .send()
         .await;
     if res.is_err() {
-        println!("Could not inform main process of sidecar initialization");
+        error!("Could not inform main process of sidecar initialization");
         std::process::exit(0);
     }
     // Init NVML
@@ -74,7 +112,7 @@ async fn main() {
     watch_main_process(main_pid);
     // Keep the HTTP server alive
     if let Err(e) = server.await {
-        eprintln!("server error: {}", e);
+        error!("Server error: {}", e);
     }
 }
 
@@ -84,7 +122,7 @@ fn watch_main_process(main_pid: u32) {
     std::thread::spawn(move || loop {
         s.refresh_processes();
         if s.process(pid).is_none() {
-            println!("Main process has exited. Stopping elevated sidecar.");
+            info!("Main process has exited. Stopping elevated sidecar.");
             std::process::exit(0);
         }
         std::thread::sleep(Duration::from_secs(1));