+++ src/app/components/main-status-bar/main-status-bar.component.ts
@@ -3,6 +3,8 @@ import { SleepService } from '../../services/sleep.service';
 import { VRChatService } from '../../services/vrchat.service';
 import { UserStatus } from 'vrchat/dist';
 import { hshrink, noop } from '../../utils/animations';
+import { firstValueFrom, Observable } from 'rxjs';
+import { OpenVRService, OpenVRStatus } from '../../services/openvr.service';
 
 @Component({
   selector: 'app-main-status-bar',
@@ -14,7 +16,7 @@ export class MainStatusBarComponent implements OnInit {
   sleepMode = this.sleepService.mode;
   user = this.vrchat.user;
 
-  constructor(private sleepService: SleepService, private vrchat: VRChatService) {}
+  constructor(private sleepService: SleepService, private vrchat: VRChatService, protected openvr: OpenVRService) {}
 
   ngOnInit(): void {}
 
@@ -30,6 +32,14 @@ export class MainStatusBarComponent implements OnInit {
         return 'var(--color-vrchat-status-red)';
       case UserStatus.Offline:
         return 'black';
+    }
+  }
+
+  async toggleSleepMode() {
+    if (await firstValueFrom(this.sleepService.mode)) {
+      await this.sleepService.disableSleepMode({ type: 'MANUAL' });
+    } else {
+      await this.sleepService.enableSleepMode({ type: 'MANUAL' });
     }
   }
 }