+++ src/app/models/settings.ts
@@ -12,3 +12,14 @@ export const APP_SETTINGS_DEFAULT: AppSettings = {
   lighthouseConsolePath:
     'C:\\Program Files (x86)\\Steam\\steamapps\\common\\SteamVR\\tools\\lighthouse\\bin\\win64\\lighthouse_console.exe',
 };
+
+export type ExecutableReferenceStatus =
+  | 'UNKNOWN'
+  | 'CHECKING'
+  | 'NOT_FOUND'
+  | 'INVALID_EXECUTABLE'
+  | 'PERMISSION_DENIED'
+  | 'INVALID_FILENAME'
+  | 'INVALID_SIGNATURE'
+  | 'UNKNOWN_ERROR'
+  | 'SUCCESS';