+++ src/app/services/lighthouse.service.ts
@@ -1,18 +1,19 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject, firstValueFrom, Observable, pairwise, startWith } from 'rxjs';
-import { LighthouseConsoleStatus, OpenVRService } from './openvr.service';
+import { OpenVRService } from './openvr.service';
 import { AppSettingsService } from './app-settings.service';
 import { invoke } from '@tauri-apps/api/tauri';
 import { OVRDevice } from '../models/ovr-device';
 import { info } from 'tauri-plugin-log-api';
+import { ExecutableReferenceStatus } from '../models/settings';
 
 @Injectable({
   providedIn: 'root',
 })
 export class LighthouseService {
-  private _consoleStatus: BehaviorSubject<LighthouseConsoleStatus> =
-    new BehaviorSubject<LighthouseConsoleStatus>('UNKNOWN');
-  public consoleStatus: Observable<LighthouseConsoleStatus> = this._consoleStatus.asObservable();
+  private _consoleStatus: BehaviorSubject<ExecutableReferenceStatus> =
+    new BehaviorSubject<ExecutableReferenceStatus>('UNKNOWN');
+  public consoleStatus: Observable<ExecutableReferenceStatus> = this._consoleStatus.asObservable();
 
   constructor(private appSettings: AppSettingsService, private openvr: OpenVRService) {
     this.init();
@@ -52,7 +53,7 @@ export class LighthouseService {
         typeof e === 'string' &&
         ['NOT_FOUND', 'PERMISSION_DENIED', 'INVALID_FILENAME'].includes(e)
       ) {
-        this._consoleStatus.next(e as LighthouseConsoleStatus);
+        this._consoleStatus.next(e as ExecutableReferenceStatus);
         return;
       }
       this._consoleStatus.next('UNKNOWN_ERROR');