+++ src/app/views/dashboard-view/views/status-automations-view/status-automations-view.component.ts
@@ -100,6 +100,7 @@ export class StatusAutomationsViewComponent implements OnInit, OnDestroy {
       )
       .subscribe();
   }
+
   ngOnDestroy() {
     this.destroy$.next();
   }